+++ src/commons/storage/mod.rs
@@ -1,6 +1,7 @@
 //! Persistent storage of data.
 
-pub use self::store::{KeyValueStore, Transaction, Error, StoreNewError};
+pub use self::backends::{Backend, Transaction, Error};
+pub use self::store::{KeyValueStore, KeyValueError};
 pub use self::types::{
     Key, Namespace, NamespaceBuf, ParseNamespaceError, ParseSegmentError,
     Scope, Segment, SegmentBuf
