+++ src/daemon/config.rs
@@ -25,8 +25,7 @@ use crate::{
         api::{PublicationServerUris, Token},
         crypto::{OpenSslSignerConfig, SignSupport},
         error::{Error, KrillIoError},
-        eventsourcing::KeyValueStore,
-        storage::Namespace,
+        storage::{KeyValueStore, Namespace},
         util::ext_serde,
         KrillResult,
     },
