+++ src/upgrades/data_migration.rs
@@ -12,9 +12,9 @@ use crate::{
             OpenSslSigner,
         },
         eventsourcing::{
-            Aggregate, AggregateStore, KeyValueStore, WalStore, WalSupport,
+            Aggregate, AggregateStore, WalStore, WalSupport,
         },
-        storage::{Namespace, Scope},
+        storage::{KeyValueStore, Namespace, Scope},
     },
     constants::{
         CASERVER_NS, KEYS_NS, PROPERTIES_NS, PUBSERVER_CONTENT_NS,
