+++ src/upgrades/pre_0_10_0/cas_migration.rs
@@ -11,10 +11,8 @@ use crate::upgrades::{
 use crate::{
     commons::{
         api::CertAuthStorableCommand,
-        eventsourcing::{
-            AggregateStore, KeyValueStore, SegmentExt,
-        },
-        storage::{Key, Segment},
+        eventsourcing::AggregateStore,
+        storage::{Key, KeyValueStore, Segment},
     },
     constants::{CASERVER_NS, CA_OBJECTS_NS},
     daemon::{
